Re-organize db modules a little
