+++ aquadoggo/src/db/models/utils.rs
