+++ aquadoggo/src/node.rs
@@ -5,7 +5,7 @@ use anyhow::Result;
 use crate::bus::ServiceMessage;
 use crate::config::Configuration;
 use crate::context::Context;
-use crate::db::sql_store::SqlStore;
+use crate::db::SqlStore;
 use crate::db::{connection_pool, create_database, run_pending_migrations, Pool};
 use crate::http::http_service;
 use crate::manager::ServiceManager;
